separate use-theme from theme-provider
